reorg nvme code.
